requirements.txt
python-telegram-bot==21.6
pandas>=2.2
requests>=2.0
openpyxl>=3.1
XlsxWriter>=3.2
python-dotenv>=0.21